Documentation Update
Added the technical Design Manual (It is behind by 1 update at the
moment) and the UML file as well that needs to opened and edited with
the free software named DIA.  The UML is behind by more commits than the
design manual though. I hope the design manual is not too messy, as I've
not had much experience with Android development and may have
misunderstood some things. Things that I simply had no idea on how to
document can be found by ctrl + f, and searching for "EDIT LATER".